Refactor httpauth
This commit refactors the internal package httpauth.

The auth file is renamed to nonce as Auth struct and its funcs are removed as they are no longer used.

Only Nonce and it's funcs remain hence the name of the file is changed to nonce.
